Remove duplicate VolumeSnapshot implementation, keep CRD-only version
- Remove src/Kinds/K8sVolumeSnapshot.php as VolumeSnapshot is not a core API
- Remove VolumeSnapshot references from InitializesResources trait
- Remove VolumeSnapshot PHPDoc annotations from KubernetesCluster
- Update all tests to use CRD registration pattern with VolumeSnapshot::register()
- Add missing methods to CRD VolumeSnapshot class for complete functionality
- Update integration tests to focus on CRD registration and basic functionality
- Update live testing script to reflect CRD-only implementation
