+++ src/KubernetesCluster.php
@@ -110,10 +110,6 @@ use RenokiCo\PhpK8s\Kinds\K8sResource;
  * @method \RenokiCo\PhpK8s\Kinds\K8sEndpointSlice getEndpointSliceByName(string $name, string $namespace = 'default', array $query = ['pretty' => 1])
  * @method \RenokiCo\PhpK8s\ResourcesList getAllEndpointSlicesFromAllNamespaces(array $query = ['pretty' => 1])
  * @method \RenokiCo\PhpK8s\ResourcesList getAllEndpointSlices(string $namespace = 'default', array $query = ['pretty' => 1])
- * @method \RenokiCo\PhpK8s\Kinds\K8sVolumeSnapshot volumeSnapshot(array $attributes = [])
- * @method \RenokiCo\PhpK8s\Kinds\K8sVolumeSnapshot getVolumeSnapshotByName(string $name, string $namespace = 'default', array $query = ['pretty' => 1])
- * @method \RenokiCo\PhpK8s\ResourcesList getAllVolumeSnapshotsFromAllNamespaces(array $query = ['pretty' => 1])
- * @method \RenokiCo\PhpK8s\ResourcesList getAllVolumeSnapshots(string $namespace = 'default', array $query = ['pretty' => 1])
  * @method \RenokiCo\PhpK8s\Kinds\K8sResource|array[\RenokiCo\PhpK8s\Kinds\K8sResource] fromYaml(string $yaml)
  * @method \RenokiCo\PhpK8s\Kinds\K8sResource|array[\RenokiCo\PhpK8s\Kinds\K8sResource] fromYamlFile(string $path, \Closure $callback = null)
  * @method \RenokiCo\PhpK8s\Kinds\K8sResource|array[\RenokiCo\PhpK8s\Kinds\K8sResource] fromTemplatedYamlFile(string $path, array $replace, \Closure $callback = null)
